Update shoulda to version 3.6.0
This change updates shoulda-matchers to version 3.1.2, and shoulda-context to
version 1.2.2.

Changes for matchers - https://github.com/thoughtbot/shoulda-matchers/blob/master/NEWS.md#312

After version 3.0 matchers no longer need `require: false` in the Gemfile, and
can just be included like a regular gem. As a side effect of that change, the
library needs an explicit config block which loads only the portions relevant to
a particular test suite (in our case, minitest and rails).

There were two tests which were erroneously checking for a `flash.now` value,
when their controllers actually just set the `flash`. This was passing on prior
shoulda versions, but now (correctly) fails, so the test was changed to match
the code.